apps/scoobi/requirements.txt
argon2-cffi==21.1.0
asttokens==2.0.5
attrs==21.2.0
autopep8==1.5.7
backcall==0.2.0
bleach==4.1.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
colorama==0.4.4
colorcet==2.0.6
coloredlogs==15.0.1
data-science-types==0.2.23
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
entrypoints==0.3
executing==0.8.0
google-api-core==2.0.0
google-auth==2.0.1
google-auth-oauthlib==0.4.6
google-cloud-bigquery==2.25.1
google-cloud-bigquery-storage==2.7.0
google-cloud-core==2.0.0
google-crc32c==1.1.3
google-resumable-media==2.0.0
googleapis-common-protos==1.53.0
grpcio==1.39.0
gspread==4.0.1
humanfriendly==9.2
icecream==2.1.1
idna==3.2
iniconfig==1.1.1
ipykernel==6.3.1
ipysankeywidget==0.4.1
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.1
jsonschema==4.0.1
jupyter-client==7.0.2
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kaleido==0.2.1
libcst==0.3.20
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
mypy==0.910
mypy-extensions==0.4.3
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.4
numpy==1.21.2
oauthlib==3.1.1
packaging==21.0
pandas==1.3.2
pandas-gbq==0.15.0
pandocfilters==1.5.0
panel==0.12.4
param==1.11.1
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
plotly==5.3.1
pluggy==1.0.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
proto-plus==1.19.0
protobuf==3.17.3
ptyprocess==0.7.0
py==1.10.0
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyct==0.4.8
pydata-google-auth==1.2.0
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.5
python-dateutil==2.8.2
pytz==2021.1
pyviz-comms==2.1.0
PyYAML==5.4.1
pyzmq==22.2.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
Send2Trash==1.8.0
six==1.16.0
tenacity==8.0.1
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
types-setuptools==57.0.2
types-toml==0.1.5
typing-extensions==3.10.0.2
typing-inspect==0.7.1
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
